separated the original codec driver into a codec regmap and tsc touchscreen/cpad driver
